+++ src/components/experienceComponents/ExperienceEdit.js
@@ -4,21 +4,21 @@ import InputItem from "../helperComponents/InputItem";
 import handleEditModeFun from "../../helperFunctions/handleEditMode";
 
 class ExperienceEdit extends React.Component {
-  handleSubmit = (contextMain, createNewExperienceItem) => {
-    handleEditModeFun(contextMain, false);
+  handleSubmit = (context, createNewExperienceItem) => {
+    handleEditModeFun(context, false);
     createNewExperienceItem();
   };
 
   render() {
-    const { contextMain, contextDisplay, createNewExperienceItem } = this.props;
+    const { context, createNewExperienceItem } = this.props;
 
     return (
       <form
         className="experience-main-container"
-        onSubmit={() => this.handleSubmit(contextMain, createNewExperienceItem)}
+        onSubmit={() => this.handleSubmit(context, createNewExperienceItem)}
       >
-        <InputItem inputType="company" type="text" context={contextDisplay} />
-        <InputItem inputType="position" type="text" context={contextDisplay} />
+        <InputItem inputType="company" type="text" context={context} />
+        <InputItem inputType="position" type="text" context={context} />
         <SubmitButton />
       </form>
     );
