+++ src/components/helperComponents/AddButton.js
@@ -3,12 +3,12 @@ import React from "react";
 // sets handleEditMode boolean from passed class to true
 class AddButton extends React.Component {
   render() {
-    const context = this.props.context;
+    const contextMain = this.props.contextMain;
 
     return (
       <button
         className="add-button"
-        onClick={() => this.props.handleEditMode(context, true)}
+        onClick={() => this.props.handleEditMode(contextMain, true)}
       >
         Add
       </button>
